+++ shared/ab-tests.js
@@ -1,0 +1,11 @@
+const COOKIE_NAME = 'ab-tests';
+
+const tests = {
+  'Just-A-Test': {
+    niceone: { weight: 0.5, value: 'nice one!' },
+    goodjob: { weight: 0.5, value: 'good job!' },
+    toobad: { weight: 4, value: 'too bad' },
+  },
+};
+
+module.exports = { COOKIE_NAME, tests };
