+++ src/state/ab-tests.js
@@ -1,12 +1,14 @@
+import { COOKIE_NAME, tests } from 'Shared/ab-tests';
 import { useGlobals } from 'State/globals';
 
 const useTest = (name) => {
   const { AB_TESTS } = useGlobals();
-  return AB_TESTS[name];
+  const assignment = AB_TESTS[name];
+  return tests[name][assignment].value;
 };
 
 const resetTests = () => {
-  document.cookie = `ab-tests=; expires=${new Date()}`;
+  document.cookie = `${COOKIE_NAME}=; expires=${new Date()}`;
 };
 
 export default useTest;
